modify review rating type to Double; add StartComponent for rating display; update related DTOs and SQL migration

+++ backend/src/main/java/com/huy/backend/models/Review.java
@@ -28,7 +28,7 @@ public class Review {
     private Movie movie;
 
     @Column(nullable = false)
-    private int rating;
+    private Double rating;
 
     @Column(length = 2000)
     private String comment;
@@ -38,5 +38,5 @@ public class Review {
 
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    
+
 }
